tripmind-ai/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic>=2.7.4
requests
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai
langchain-community
langchain-text-splitters
langgraph
langsmith
chromadb
tavily-python
pypdf
google-generativeai